lab01/pyproject.toml
[project]
name = "lab01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "python-dotenv>=1.1.1",
]



[tool.ruff]
line-length = 140
